move SignalementDraftCleaner to specific class #5748
